+++ sdk/highlight-hono/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @highlight-run/hono
 
+## 1.0.25
+
+### Patch Changes
+
+- @highlight-run/node@3.12.22
+
 ## 1.0.24
 
 ### Patch Changes