+++ sdk/highlight-node/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @highlight-run/node
 
+## 3.12.22
+
+### Patch Changes
+
+- Updated dependencies [7e8f9e5]
+- Updated dependencies [7e8f9e5]
+    - highlight.run@9.20.0
+
 ## 3.12.21
 
 ### Patch Changes